g2h: Fix white screen before logo
Currently we notice that quick white screen appears prior to the
logo, which is a bit unpleasant.

Fix this by keeping the display turned off initially and only turn it
on at a later stage, inside board_late_init().

Experiments show that a 150ms delay needs to be added prior
to toggling LVDS0_EN.

Tested on a G2H board with a parallel 7" display.

Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>